Add City Directory: 1886

+++ clean-build.sh
@@ -5,9 +5,11 @@ rm -v _cmhc/*
 rm -v _directories/*
 rm -v _census/*
 rm -v search/*
-mv _site .trash
-rm -rf .trash &
 
+if [ -z _site ]; then
+  mv _site .trash
+  rm -rf .trash &
+fi
 
 #bundle exec rake wax:derivatives:iiif cmhc
 #bundle exec rake wax:derivatives:iiif directories
@@ -22,8 +24,8 @@ bundle exec rake wax:search census
 #read -p "continue with 'clean' when ready..." -n 1
 date; time bundle exec jekyll clean
 
-read -p "continue with 'build' when ready..." -n 1
-date; time JEKYLL_ENV=production bundle exec jekyll build --profile > build.out 2>&1
+#read -p "continue with 'build' when ready..." -n 1
+date; time JEKYLL_ENV=production bundle exec jekyll build --profile > misc/build.out 2>&1
 
 #date; time aws s3 sync _site/ s3://history.lakecountypubliclibrary.org/cmhc/ --delete --exclude clean-build.sh --exclude notes.txt --exclude s3_website.yml --exclude misc --size-only --profile lcpl
 #date; time aws s3 sync --dryrun _site/ s3://history.lakecountypubliclibrary.org/cmhc/ --delete --exclude clean-build.sh --exclude notes.txt --exclude s3_website.yml --exclude misc --size-only --profile lcpl
